# Hestia-Swagger/swagger.yml
openapi: 3.0.3
info:
  title: Hestia — API Reference
  description: |
    Aggregated API reference for all Hestia services.
    Each path overrides the server URL to route to the correct service.

    | Service   | Port  |
    |-----------|-------|
    | Hub       | 19001 |
    | Archive   | 19002 |
    | Hecate    | 19003 |
    | Oracle    | 19004 |
    | Hermes    | 19005 |
    | Scout     | 19006 |
    | Chronos   | 19007 |
    | Argus     | 19008 |
    | Iris      | 19012 |
    | Hephaestus| 19010 |
    | Dummy     | 19011 |
  version: 1.0.0
servers:
  - url: http://localhost:19001
    description: Hub (default)
...
components:
  schemas:
...
    ArgusRemediationRequest:
      type: object
      required: [service, issue]
      properties:
        service:
          type: string
        issue:
          type: string
        severity:
          type: string
          default: warning
        requested_action:
          type: string
          default: runbook_autoselect
        environment:
          type: string
          default: dev
        dry_run:
          type: boolean
          default: true
        auto_approve:
          type: boolean
          default: false
        metadata:
          type: object
          additionalProperties: true
...
paths:
...
  /api/logs:
    get:
      tags: [Hub]
      summary: Retrieve buffered runtime logs (service-local endpoint)
      description: |
        Implemented in all runtime services.
        Current instances:
        - Hub (19001)
        - Archive (19002)
        - Hecate (19003)
        - Oracle (19004)
        - Hermes (19005)
        - Scout (19006)
        - Chronos (19007)
        - Argus (19008)
        - Athena (19009)
        - Hephaestus (19010)
        - Telegram control API (internal: hestia_telegram:8010)

        Sensitive values (tokens, passwords, secrets, credential URLs, auth headers)
        are redacted before entries are buffered and returned.

        Use query filters to keep output dense and useful for troubleshooting.
      servers:
        - url: http://localhost:19001
          description: Hub
        - url: http://localhost:19002
          description: Archive
        - url: http://localhost:19003
          description: Hecate
        - url: http://localhost:19012
          description: Iris
        - url: http://localhost:19004
          description: Oracle
        - url: http://localhost:19005
          description: Hermes
        - url: http://localhost:19006
          description: Scout
        - url: http://localhost:19007
          description: Chronos
        - url: http://localhost:19008
          description: Argus
        - url: http://localhost:19009
          description: Athena
        - url: http://localhost:19010
          description: Hephaestus
        - url: http://hestia_telegram:8010
          description: Telegram control API (Docker network)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
            maximum: 2000
        - name: level
          in: query
          schema:
            type: string
            example: INFO
        - name: contains
          in: query
          schema:
            type: string
            example: outbound_event_id=p2-4-smoke-3
      responses:
        "200":
          description: "{service, count, logs[]}"
...
  /api/argus/logs:
    get:
      tags: [Argus]
      summary: Query collected service logs
      servers:
        - url: http://localhost:19008
      parameters:
        - name: service
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
            default: WARNING
        - name: since
          in: query
          schema:
            type: string
            default: 30m
          description: "e.g. '30m', '2h', '1d'"
      responses:
        "200":
          description: "{count, filters{}, events[]}"
...
  /api/argus/remediate:
    post:
      tags: [Argus]
      summary: Forward remediation intent to Hephaestus through Hub routing
      servers:
        - url: http://localhost:19008
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArgusRemediationRequest"
      responses:
        "200":
          description: "{status, forwarded_to, result}"

  # ══════════════════════════════════════════════════════════════════════════
  #  HEPHAESTUS — 19010
  # ══════════════════════════════════════════════════════════════════════════
...